+++ src/App.js
@@ -1,7 +1,7 @@
 import Header from './components/header/header';
 import Main from './components/main/main';
-// import MessageStatus from './components/send-history/send-history';
 import '../src/scss/app.scss';
+// import MessageStatus from './components/message-status/message-status';
 
 function App() {
   return (
